feat(hypr, yazi, script): improve config
fix(hypr): duplicate bind
refactor(yazi): go to Documents instead of Downloads
refactor(submap): remove redundant ESCAPE submap reset
feat(mouse-click): add mouse toggle copy and state management
refactor(mouse_state): use if-elif and improve state handling
fix(hypr): quoting on vars and add new submap
refactor(hypr): rename config files for clarity and consistency
refactor(hyprsunset): simplify profile configuration
feat(windowrule): add new workspace rules

+++ scripts/mouse-click.sh
@@ -2,22 +2,56 @@
 set -euo pipefail
 
 STATE_FILE="/tmp/mouse_state"
-ydotool click 0x80 2>/dev/null || true
-echo "released" >"$STATE_FILE"
 
-case "$1" in
-left) ydotool click 0x40 0x80 ;;
-right) ydotool click 0x41 0x81 ;;
-middle) ydotool click 0x42 0x82 ;;
-toggle)
+# Initialize state file if missing
+if [[ ! -f "$STATE_FILE" ]]; then
+  echo "released" >"$STATE_FILE"
+fi
+
+if [[ "$1" = "left" ]]; then
+  ydotool click 0x40 0x80
+
+elif [[ "$1" = "right" ]]; then
+  ydotool click 0x41 0x81
+
+elif [[ "$1" = "middle" ]]; then
+  ydotool click 0x42 0x82
+
+elif [[ "$1" = "toggle" ]]; then
   CURRENT_STATE=$(cat "$STATE_FILE")
 
-  if [ "$CURRENT_STATE" = "released" ]; then
+  if [[ "$CURRENT_STATE" = "released" ]]; then
+    # Start Pressing
     echo "pressed" >"$STATE_FILE"
     ydotool click 0x40
+
   else
+    # End Pressing & Copy
     echo "released" >"$STATE_FILE"
     ydotool click 0x80
+
+    sleep 0.1
+    # Ctrl+C (copy selected text)
+    ydotool key 29:1 46:1 46:0 29:0
+    sleep 0.1
+
+    # Click (clear selection)
+    ydotool click 0x40 0x80
+
+    notify-send -h boolean:transient:true \
+      "Copied." \
+      -i dialog-information \
+      -t 1400
+
+    hyprctl dispatch submap reset
   fi
-  ;;
-esac
+
+elif [[ "$1" = "reset" ]]; then
+  echo "released" >"$STATE_FILE"
+
+  ydotool click 0x80
+  # ydotool click 0x40 0x80
+  # ydotool key 1:1 1:0
+
+  hyprctl dispatch submap reset
+fi
